requirements.txt
flasgger
mailjet_rest
PyJWT
alembic==0.9.5
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.2
Flask-Script==2.0.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
psycopg2==2.7.3.1
Werkzeug==0.12.2
